Add FileController & move upload method
