+++ src/main/java/com/web/yapp/server/controller/FileController.java
@@ -1,0 +1,23 @@
+package com.web.yapp.server.controller;
+
+import com.web.yapp.server.service.S3Uploader;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+
+@RequiredArgsConstructor
+@Controller
+public class FileController {
+    private final S3Uploader s3Uploader;
+
+    @PostMapping("/upload")
+    @ResponseBody
+    public String upload(@RequestParam("data") MultipartFile multipartFile) throws IOException {
+        return s3Uploader.upload(multipartFile, "static");
+    }
+}
