+++ src/main/resources/templates/index.mustache
@@ -11,4 +11,6 @@
     </div>
 </div>
 <br>
+
+
 {{>layout/footer}}
